// angularProject/src/app/api/model/orderDto.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderStatus } from './orderStatus';
import { PaymentStatus } from './paymentStatus';


export interface OrderDto { 
    id?: number;
    keyUsage?: string | null;
    amountPaid?: number | null;
    symbol?: string | null;
    customerId?: number | null;
    stockTipId?: number | null;
    orderGuid?: string;
    storeId?: number;
    billingAddressId?: number;
    orderStatusId?: number;
    paymentStatusId?: number;
    paymentMethodSystemName?: string | null;
    customerCurrencyCode?: string | null;
    currencyRate?: number;
    orderSubTotalDiscountInclTax?: number;
    orderSubTotalDiscountExclTax?: number;
    orderDiscount?: number;
    orderTotal?: number;
    paidDateUtc?: string | null;
    createdOnUtc?: string;
    customOrderNumber?: string | null;
    redeemedCoinsEntryId?: number | null;
    productType?: string | null;
    paymentType?: string | null;
    advisorId?: number;
    orderStatus?: OrderStatus;
    paymentStatus?: PaymentStatus;
}
export namespace OrderDto {
}