// angularProject/src/app/api/model/customerBasicInfoSchema.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CustomerBasicInfo } from './customerBasicInfo';


export interface CustomerBasicInfoSchema { 
    customerBasicInfo?: CustomerBasicInfo;
    isSuccess?: boolean;
    message?: string | null;
}